+++ CHANGELOG.md
@@ -13,6 +13,12 @@
 **Note**: Gaps between patch versions are faulty/broken releases. **Note**: A feature tagged as Experimental is in a
 high state of flux, you're at risk of it changing without notice.
 
+# 0.5.7
+
+- **New Feature**
+  - add `readonlyNonEmptyArray` (@gcanti)
+  - add `readonlySetFromArray` (@gcanti)
+
 # 0.5.6
 
 - **Polish**